+++ tests/test_export.py
@@ -306,6 +306,18 @@ def test_export_requirements(tmp_empty, code, expected):
     assert Path('requirements.txt').read_text() == expected
 
 
+def test_export_requirements_doesnt_overwrite(tmp_empty):
+    reqs = Path('requirements.txt')
+    reqs.write_text('soorgeon\n')
+
+    exporter = export.NotebookExporter(_read(definition_with_import))
+    exporter.export_requirements()
+
+    expected = ('soorgeon\n# Auto-generated file, may need manual '
+                'editing\nploomber\nmatplotlib\n')
+    assert reqs.read_text() == expected
+
+
 def test_does_not_create_exported_py_if_no_definitions(tmp_empty):
     exporter = export.NotebookExporter(_read(simple))
     exporter.export_definitions()